+++ tests/client_test_cases/get_microgrid/defaults_case.py
@@ -7,8 +7,8 @@ from datetime import datetime, timezone
 from typing import Any
 from unittest.mock import AsyncMock
 
-from frequenz.api.assets.v1 import assets_pb2
 from frequenz.api.common.v1alpha8.microgrid import microgrid_pb2
+from frequenz.api.platformassets.v1alpha1 import platformassets_pb2 as assets_pb2
 from frequenz.client.common.microgrid import EnterpriseId, MicrogridId
 
 from frequenz.client.assets import Microgrid, MicrogridStatus